Item9715: Can't call DBNameSafe. on unblessed ref
For some reason I was getting this on one system and not another when
trying to do an initial import. On line 190 MongoDBPlugin.pm. Seems the
$Foswiki::Plugins::SESSION->{MongODB} was just a hashref containing only
a skeleton of an timing info thingy, sorry I couldn't c&p more detail.

So, check the ref is blessed, rather than merely defined.

git-svn-id: http://svn.foswiki.org/trunk/MongoDBPlugin@15922 0b4bb1d4-4e5a-0410-9cc4-b2b747904278